# shard buffers to reduce lock contention