--- a/statsd/options_test.go
+++ b/statsd/options_test.go
@@ -17,6 +17,7 @@ func TestDefaultOptions(t *testing.T) {
 	assert.Equal(t, options.MaxMessagesPerPayload, DefaultMaxMessagesPerPayload)
 	assert.Equal(t, options.BufferPoolSize, DefaultBufferPoolSize)
 	assert.Equal(t, options.BufferFlushInterval, DefaultBufferFlushInterval)
+	assert.Equal(t, options.BufferShardCount, DefaultBufferShardCount)
 	assert.Equal(t, options.SenderQueueSize, DefaultSenderQueueSize)
 	assert.Equal(t, options.WriteTimeoutUDS, DefaultWriteTimeoutUDS)
 	assert.Equal(t, options.Telemetry, DefaultTelemetry)
@@ -29,6 +30,7 @@ func TestOptions(t *testing.T) {
 	testMaxMessagePerPayload := 1024
 	testBufferPoolSize := 32
 	testBufferFlushInterval := 48 * time.Second
+	testBufferShardCount := 28
 	testSenderQueueSize := 64
 	testWriteTimeoutUDS := 1 * time.Minute
 
@@ -39,6 +41,7 @@ func TestOptions(t *testing.T) {
 		WithMaxMessagesPerPayload(testMaxMessagePerPayload),
 		WithBufferPoolSize(testBufferPoolSize),
 		WithBufferFlushInterval(testBufferFlushInterval),
+		WithBufferShardCount(testBufferShardCount),
 		WithSenderQueueSize(testSenderQueueSize),
 		WithWriteTimeoutUDS(testWriteTimeoutUDS),
 		WithoutTelemetry(),
@@ -51,6 +54,7 @@ func TestOptions(t *testing.T) {
 	assert.Equal(t, options.MaxMessagesPerPayload, testMaxMessagePerPayload)
 	assert.Equal(t, options.BufferPoolSize, testBufferPoolSize)
 	assert.Equal(t, options.BufferFlushInterval, testBufferFlushInterval)
+	assert.Equal(t, options.BufferShardCount, testBufferShardCount)
 	assert.Equal(t, options.SenderQueueSize, testSenderQueueSize)
 	assert.Equal(t, options.WriteTimeoutUDS, testWriteTimeoutUDS)
 	assert.Equal(t, options.Telemetry, false)