### What does this PR do?

We had a single buffer we wrote all serialized metrics to. It was creating contention when multiple Goroutines where pushing metrics at the same time. This PR tries to address this issue by instead using a configurable number of buffers to reduce the chance two Goroutines write to the same buffer at the same time.
To implement this, we need a way to decide to which buffer we should write when a metric is submitted.
Currently, this PR uses the metric name to do so:
https://github.com/DataDog/datadog-go/blob/12805b4c7a2528b930591bd6d14fea0e031d6aa2/statsd/statsd.go#L389-L390

### Benchmarks

Using `bench.RunParallel`. Each goroutine sends metrics with a different name:
```
benchmark                      old ns/op     new ns/op     delta
BenchmarkStatsdUDP-4          219           105           -52.05%
BenchmarkStatsdUDS-4          217           110           -49.31%
```

### Notes

- Sharding by metric name might not be optimal is some scenarios.
- Buffering might get a bit worse as some of the buffers might get rarely written to which would lead to datagrams containing few metrics.

### TODO

- [ ] add MIT licence & copyright to fnv1a || import it as a dependency