@kevinconaway 2019-12-18
Thoughts on allowing clients to control the "sharding"?  Perhaps something like

```go
client := statsd.Client

for range workerCount {
   go func() {
      stats := client.NewClient()
      //....
   }()
}
```

Each "sub" client would then be independent from the others and could buffer / append metrics internally before sending them to to be delivered.

I'm concerned that hashing every metric key is going to add unnecessary overhead for the scenarios where we submit metrics in tight loops