@arbll 2019-12-18
@kevinconaway I could add a fast path for the case where we only configured one shard. That should work for such use cases right ?